Included <setjmp.h> in order to perform proper error handling Currently the program aborts in case it finds an error when encoding or decoding an image.
Having a module closing your program is not error handling. This would complicate matters when adding the ability to transcode the image or adding another interface.

So the idea here is to have the caller setting a long jump to itself, then the jump is performed on error, and the caller handles the error as it see fit.