Simplified error handling
My goal is just to prevent the image libraries from closing the program on error.

The error handling strategy I was going for previously was different than for other errors, on the top making things more complex for the program (it would be necessary to set a couple of long jumps for just this scenario).

So I found simpler to "glue" this part with the rest of the program by just have a single error code for encode/decode errors, and have a global variable with a string detailing the error.

This is more or less similar than the standard `strerror(errno)`, and should work in our case since the program does not need to concurrently handle multiple cover images at once.