Refactored the swapping of elements on an array
This does not change the output, it was made for clarity's sake. When I first wrote that code it didn't occur to me that XORing would add some unnecessary operations, since the temp variable will just be on a CPU register.

It is worth noting that the old code didn't affect the performance, since with optimizations enabled the compiler already replaced the XOR swap by a direct swap. Using a temporary variable now is a matter of clarity.