The index of the PRNG buffer is now checked if it is greater than or equal to the buffer size, instead of if it is just equal
When the program is working correctly it should never be greater than the buffer size, since the index is always incremented by one.

However it is better to be safe than sorry, checking if it is also greater than should prevent or minimize the damage in case a buffer overflow arises on the future. No body wants a ship to sink, but it is still nice having enough lifeboats around :)