Fixed check if the WebP muxers were created successfully
That semicolon on the end of the if statement was an oversight... I dont't know how I let that pass, but fortunately it seems to be rather unlikely that muxer creation to fail as it does not use too much memory. Also the libwebp's Mux functions check for NULL anyways, so no harm was done.