Removed static functions from the header files
Because they are not meant to be included on all source files of the project. Their prototypes were moved to the source files where the functions are going to be used.

This should help the code linter, since functions aren't going to be suggested on files where they can't be used.